Projections: drop redundant .astype(F32) in matmul hot path
* The forwarder's projection matmul chain was wrapping each
  weight in .astype(np.float32) before the matmul. The
  default for astype is copy=True, so even when the
  dtype already matched the call allocated a fresh buffer
  and ran a full F32-to-F32 memcpy. On Qwen3-14B Q4_K_M this
  measured 15-100 ms per projection call (gate_w alone
  ~108 ms); 30-60 seconds of pure memcpy waste across a
  full decode pass.

* All 14 hot-path sites now use .astype(np_dtype,
  copy=False):
    - qwen2 _decoder_block: qkv, o_proj, gateup, down_proj
    - _gemma3_decoder_block: same four
    - _qwen35_full_attention_block: same four + attn gate
    - _compute_layer_logits: head_w + matmul output
  Behaviour: when dtype matches (production default
  float32) the call returns the input buffer unchanged
  in ~0.003-0.4 ms. When dtype differs (caller picks
  float16 / bfloat16) it still does the conversion.

* Defence-in-depth: _fetch_proj non-quantized branch
  and the bias fetch also use copy=False so a future
  refactor that quietly broadens dtype still won't pay
  the redundant copy on already-aligned buffers.

* Synthetic benchmark on 14 B decode shapes (seq=1):
  per-layer matmul work dropped 270 ms -> 65 ms (4.16x).
  40 layers * 5 decode tokens saves ~41 seconds.

* End-to-end on Qwen3-0.6B Q4_K_M (--max-new 10):
  top-4 projection time 10765 ms -> 7529 ms (~30%
  reduction). Dequantization share rose 14.6% -> 27.8%
  in the profile percentage but only because non-dequant
  work shrank faster than dequant; absolute dequant time
  is unchanged.

* Regression test added: test_qwen2_projection_matmul_
  does_not_astype_copy pins the contract that projection
  weights survive a forward pass with their dtype
  unchanged.

* 163 tests pass (was 162; +1 regression test).

+++ src/flatrun/model/qwen2.py
@@ -498,7 +498,7 @@ def _fetch_proj(
         if dequant_cache is not None:
             dequant_cache[cache_key] = arr
     else:
-        arr = weight_handle.as_numpy().astype(dtype)
+        arr = weight_handle.as_numpy().astype(dtype, copy=False)
     return arr
 
 
@@ -529,7 +529,7 @@ def _fetch_linear_with_quant(
     else:
         bias_name = f"model.layers.{layer_index}.{proj_name}.bias"
     if bias_name in handles:
-        bias = handles[bias_name].as_numpy().astype(np.float32)
+        bias = handles[bias_name].as_numpy().astype(np.float32, copy=False)
     else:
         bias = None
     return weight, bias
@@ -776,7 +776,7 @@ def make_qwen2_forwarder(
             normed = _gemma3_norm(h, norm_w, config.rms_norm_eps)
         else:
             normed = _rms_norm(h, norm_w, config.rms_norm_eps)
-        return (normed @ head_w.astype(np.float32).T).astype(np.float32)
+        return (normed @ head_w.astype(np_dtype, copy=False).T).astype(np_dtype, copy=False)
 
     # Dequant cache is on by default. When disabled the forwarder
     # runs in pure streaming mode: every layer's weights are
@@ -1157,7 +1157,16 @@ def make_qwen2_forwarder(
         k_w = _as_linear(k_w, config.hidden_size, f"layer{idx}.k_proj")
         v_w = _as_linear(v_w, config.hidden_size, f"layer{idx}.v_proj")
         with _p("qkv_proj", _P):
-            qkv_w = np.concatenate([q_w, k_w, v_w], axis=0).astype(np.float32)
+            # q_w/k_w/v_w are already F32 (from ``_fetch_proj`` with the
+            # production ``dtype=float32``). The previous code did
+            # ``.astype(np.float32)`` on the concatenated buffer; the
+            # default for ``astype`` is ``copy=True`` so it allocated a
+            # 147 MB F32 buffer per decoder step on a 14 B model just
+            # to copy data that was already F32. Measuring 50-100 ms
+            # per call on Apple Silicon. ``copy=False`` returns the
+            # same buffer when the dtype matches (0.003-0.4 ms) and
+            # only pays the conversion cost when it does not.
+            qkv_w = np.concatenate([q_w, k_w, v_w], axis=0).astype(np_dtype, copy=False)
             qkv = x @ qkv_w.T
             if q_b is not None or k_b is not None or v_b is not None:
                 qkv = qkv + np.concatenate(
@@ -1253,7 +1262,11 @@ def make_qwen2_forwarder(
             )
         o_w = _as_linear(o_w, q_dim, f"layer{idx}.o_proj")
         with _p("o_proj", _P):
-            attn_out = attn_out @ o_w.astype(np.float32).T
+            # See the QKV comment above. ``o_w`` is already F32 from
+            # ``_fetch_proj``; ``astype(np_dtype, copy=False)`` is the
+            # zero-cost no-op here and only pays the conversion cost
+            # when the caller chose a non-F32 dtype.
+            attn_out = attn_out @ o_w.astype(np_dtype, copy=False).T
             if o_b is not None:
                 attn_out = attn_out + o_b
         with _p("residual_attn", _P):
@@ -1286,7 +1299,10 @@ def make_qwen2_forwarder(
         down_w = _as_linear(down_w, inter, f"layer{idx}.down_proj")
 
         with _p("gateup_proj", _P):
-            gateup = x @ np.concatenate([gate_w, up_w], axis=0).astype(np.float32).T
+            # gate_w and up_w are F32 from ``_fetch_proj``. Copy=False
+            # on the dtype alignment is zero-cost when dtype matches.
+            # Same pattern as the QKV concat above.
+            gateup = x @ np.concatenate([gate_w, up_w], axis=0).astype(np_dtype, copy=False).T
         gate = gateup[:, :inter]
         up = gateup[:, inter:]
         with _p("silu", _P):
@@ -1296,7 +1312,8 @@ def make_qwen2_forwarder(
         with _p("mul", _P):
             silu_mul = silu_gate * up
         with _p("down_proj", _P):
-            mlp_out = silu_mul @ down_w.astype(np.float32).T
+            # Same zero-cost-dtype-match pattern as QKV / O above.
+            mlp_out = silu_mul @ down_w.astype(np_dtype, copy=False).T
         with _p("residual_mlp", _P):
             out = residual + mlp_out
         # Release every per-layer weight and intermediate so the
@@ -1371,7 +1388,7 @@ def make_qwen2_forwarder(
         q_w = _as_linear(q_w, config.hidden_size, f"layer{idx}.q_proj")
         k_w = _as_linear(k_w, config.hidden_size, f"layer{idx}.k_proj")
         v_w = _as_linear(v_w, config.hidden_size, f"layer{idx}.v_proj")
-        qkv_w = np.concatenate([q_w, k_w, v_w], axis=0).astype(np.float32)
+        qkv_w = np.concatenate([q_w, k_w, v_w], axis=0).astype(np_dtype, copy=False)
         qkv = x @ qkv_w.T
 
         q = qkv[:, :q_dim].reshape(seq_len, n_heads, head_dim)
@@ -1425,7 +1442,7 @@ def make_qwen2_forwarder(
             handles, idx, "self_attn.o_proj", config, np_dtype, dequant_cache=dequant_cache
         )
         o_w = _as_linear(o_w, q_dim, f"layer{idx}.o_proj")
-        attn_out = attn_out @ o_w.astype(np.float32).T
+        attn_out = attn_out @ o_w.astype(np_dtype, copy=False).T
         hidden = residual + attn_out
 
         # ----- MLP (SwiGLU) with pre/post norms -----
@@ -1449,11 +1466,11 @@ def make_qwen2_forwarder(
         inter = gate_w.shape[0]
         down_w = _as_linear(down_w, inter, f"layer{idx}.down_proj")
 
-        gateup = x @ np.concatenate([gate_w, up_w], axis=0).astype(np.float32).T
+        gateup = x @ np.concatenate([gate_w, up_w], axis=0).astype(np_dtype, copy=False).T
         gate = gateup[:, :inter]
         up = gateup[:, inter:]
         silu_gate = gate / (1.0 + np.exp(-np.clip(gate, -80.0, 80.0)))
-        mlp_out = (silu_gate * up) @ down_w.astype(np.float32).T
+        mlp_out = (silu_gate * up) @ down_w.astype(np_dtype, copy=False).T
         hidden = residual + mlp_out
 
         # Gemma 3 applies ``post_feedforward_layernorm`` to the *whole*
@@ -1517,7 +1534,7 @@ def make_qwen2_forwarder(
         q_w = _as_linear(q_w, config.hidden_size, f"layer{idx}.q_proj")
         k_w = _as_linear(k_w, config.hidden_size, f"layer{idx}.k_proj")
         v_w = _as_linear(v_w, config.hidden_size, f"layer{idx}.v_proj")
-        qkv_w = np.concatenate([q_w, k_w, v_w], axis=0).astype(np.float32)
+        qkv_w = np.concatenate([q_w, k_w, v_w], axis=0).astype(np_dtype, copy=False)
         qkv = x @ qkv_w.T
         q = qkv[:, :q_dim].reshape(seq_len, n_heads, head_dim)
         k = qkv[:, q_dim : q_dim + kv_dim].reshape(seq_len, n_kv_heads, head_dim)
@@ -1574,14 +1591,14 @@ def make_qwen2_forwarder(
                 handles, idx, "self_attn.gate_proj", config, np_dtype, dequant_cache=dequant_cache
             )
             gate_w = _as_linear(gate_w, config.hidden_size, f"layer{idx}.gate_proj")
-            gate = (x @ gate_w.astype(np.float32).T)
+            gate = (x @ gate_w.astype(np_dtype, copy=False).T)
             attn_out = attn_out * _sigmoid(gate)
 
         o_w, _ = _fetch_linear_with_quant(
             handles, idx, "self_attn.o_proj", config, np_dtype, dequant_cache=dequant_cache
         )
         o_w = _as_linear(o_w, q_dim, f"layer{idx}.o_proj")
-        attn_out = attn_out @ o_w.astype(np.float32).T
+        attn_out = attn_out @ o_w.astype(np_dtype, copy=False).T
         hidden = residual + attn_out
 
         # MLP - same SwiGLU as Qwen2 / Qwen3. Qwen3.5's per-layer
@@ -1606,11 +1623,11 @@ def make_qwen2_forwarder(
         up_w = _as_linear(up_w, config.hidden_size, f"layer{idx}.mlp_up")
         inter = gate_w.shape[0]
         down_w = _as_linear(down_w, inter, f"layer{idx}.mlp_down")
-        gateup = x @ np.concatenate([gate_w, up_w], axis=0).astype(np.float32).T
+        gateup = x @ np.concatenate([gate_w, up_w], axis=0).astype(np_dtype, copy=False).T
         gate = gateup[:, :inter]
         up = gateup[:, inter:]
         silu_gate = gate / (1.0 + np.exp(-np.clip(gate, -80.0, 80.0)))
-        mlp_out = (silu_gate * up) @ down_w.astype(np.float32).T
+        mlp_out = (silu_gate * up) @ down_w.astype(np_dtype, copy=False).T
         out = residual + mlp_out
         del (
             q_w, k_w, v_w, qkv_w, qkv, q, k, v,
@@ -1772,7 +1789,7 @@ def make_qwen2_forwarder(
                 dequant_cache=dequant_cache,
             )
             head_w = _as_linear(head_w, config.hidden_size, "lm_head")
-        logits = (hidden @ head_w.astype(np.float32).T).astype(np.float32)
+        logits = (hidden @ head_w.astype(np_dtype, copy=False).T).astype(np_dtype, copy=False)
         if getattr(config, "debug_trace", False):
             # Same per-token table, but with logits attached so the
             # entropy and confidence columns are now populated.
